# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Dipes & Wipes API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stripe/checkout/subscription:
    post:
      operationId: createSubscriptionCheckout
      tags: [stripe]
      summary: Create a Stripe Checkout session for a subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionCheckoutInput"
      responses:
        "200":
          description: Stripe Checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSession"
        "400":
          description: Invalid input or missing Stripe price
        "404":
          description: Subscription not found
...
  /subscriptions:
    get:
      operationId: listSubscriptions
      tags: [subscriptions]
      summary: List all subscriptions
      responses:
        "200":
          description: List of subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscription"
    post:
      operationId: createSubscription
      tags: [subscriptions]
      summary: Create a subscription for a baby
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /subscriptions/{id}:
    get:
      operationId: getSubscription
      tags: [subscriptions]
      summary: Get a subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
    patch:
      operationId: updateSubscription
      tags: [subscriptions]
      summary: Update a subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
    delete:
      operationId: deleteSubscription
      tags: [subscriptions]
      summary: Cancel a subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Cancelled
  /subscriptions/{id}/pause:
    post:
      operationId: pauseSubscription
      tags: [subscriptions]
      summary: Pause a subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /subscriptions/{id}/resume:
    post:
      operationId: resumeSubscription
      tags: [subscriptions]
      summary: Resume a paused subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
...
components:
  schemas:
...
    Subscription:
      type: object
      required: [id, babyId, status, frequency, nextDeliveryDate, createdAt]
      properties:
        id:
          type: integer
        babyId:
          type: integer
        babyName:
          type: ["string", "null"]
        status:
          type: string
        frequency:
          type: string
        brand:
          type: ["string", "null"]
        currentDiaperSize:
          type: ["string", "null"]
        nextDeliveryDate:
          type: string
          format: date
        monthlyPriceCents:
          type: ["integer", "null"]
        shippingName:
          type: ["string", "null"]
        shippingAddress1:
          type: ["string", "null"]
        shippingCity:
          type: ["string", "null"]
        shippingState:
          type: ["string", "null"]
        shippingZip:
          type: ["string", "null"]
        shippingCountry:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/SubscriptionItem"
        createdAt:
          type: string
    SubscriptionItem:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        productName:
          type: ["string", "null"]
        quantity:
          type: integer
        priceCents:
          type: ["integer", "null"]
...
    SubscriptionCheckoutInput:
      type: object
      required: [subscriptionId]
      properties:
        subscriptionId:
          type: integer
        successUrl:
          type: string
        cancelUrl:
          type: string
    CheckoutSession:
      type: object
      required: [url]
      properties:
        url:
          type: string
...
    SubscriptionInput:
      type: object
      required: [babyId, frequency, nextDeliveryDate]
      properties:
        babyId:
          type: integer
        frequency:
          type: string
        nextDeliveryDate:
          type: string
          format: date
        brand:
          type: string
        shippingName:
          type: string
        shippingAddress1:
          type: string
        shippingCity:
          type: string
        shippingState:
          type: string
        shippingZip:
          type: string
        shippingCountry:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/SubscriptionItemInput"
    SubscriptionItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
    SubscriptionUpdate:
      type: object
      properties:
        frequency:
          type: string
        nextDeliveryDate:
          type: string
          format: date
        shippingName:
          type: string
        shippingAddress1:
          type: string
        shippingCity:
          type: string
        shippingState:
          type: string
        shippingZip:
          type: string
        shippingCountry:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/SubscriptionItemInput"
...